rpm83p: use asm_Flag{1,2,3} for various flags
These are provided by the TI-OS, and accessed as offsets from the IY register.
This allows the HL register to be used for other purposes. And using the IY
register is actually faster and smaller. The `bit n, (iy+offset` takes 4 bytes
and 20 clock cycles, but the combo of `ld hl, address; bit n, (hl)` takes 5
bytes and 22 (= 10 + 12) clock cyles.