prime.asm;menuhandler: PRIM: push the prime factor onto stack, retaining the original X
The prime factor is pushed onto the RPN stack, retaining the original X, which
becomes Y. This allows the user to hit '/', then `PRIM` again to get next
candidate prime factor. If this process is iterated until a '1' is returned,
all prime factors of the original X can be found.