Split into a multi-page flash app
RPN83P is now larger than one flash page (16 kiB), so we are forced to create a
multi-page flash application. The first chuck of code to move to Page 1 is the
HELP menuand its related handlers and string content. That moves about 1 kB to
Page 1.

This refactoring forces the main parsing loop (processCommand()) to use the
'handlerCode' return code to detect the 2ND QUIT command, instead of calling
the mainExit() function directly from the HELP parsing (processHelp()). The
mainExit() function resides on Page 0, so it is inaccessible from processHelp()
on Page 1.

Once we use the 'handlerCode' for AppQuit, it is cleaner to use the same
mechanism to handle the "help exit" event, which needs to clear the screen
before returning to the main parsing loop.