rpn83p: support jumpBack from 'MODE' shortcut
Normally, ON/EXIT goes up to the parent MenuGroup of the current MenuGroup.
However, when the MODE button is pressed, it acts as a direct shortcut to the
`ROOT > MODE` from anywhere in the menu hierarchy. In this case, it seems more
convenient for the ON/EXIT button to go back to the original MenuGroup, instead
of going up to the parent of the `ROOT > MODE` MenuGroup (which would be the
ROOT).

Currently, the jumpBack menu is implemented *only* for the MODE button. The
STAT button would be another candidate, but the jumpBack menu is *not*
implemented for the STAT shortcut because it seems likely that the user would
spend a considerable amount of time underneath the `ROOT > STAT` MenuGroup, and
it seems more useful for the ON/EXIT to do the normal action.