menuhandlers: flip order of (X,Y) and (r,theta) for polar-rectangular conversion
Now the `X` register holds the `x` or `r` value, and the `Y` register holds the
`y` or `theta` value. This makes the `>POL` and `>REC` function consistent with
the equivalent functions on the HP-42S. I'm not sure why I had them reversed,
maybe it's because that's how the TI-OS ordered these parameters. Also add new
Help page named 'CONV Arguments' to summarize the order of (x,y,r,theta)
parameters.