display.asm,rpn83p.asm: use rpnFlagsShowModeEnabled flag for SHOW mode
The previous implementation used the 'drawMode' variable and a new
'drawModeShow' to implement the SHOW mode. But it occurred to me that this is
dangerous because the drawMode variable is user-modifiable. If the user selects
'drawModeShow' explicitly, the app would be stuck in 'drawModeShow' forever,
because the prompt for '2ND DRAW' cannot be rendered in SHOW mode, so there
would be no way that the user can change the current draw mode back to normal
mode. Using the new 'rpnFlagsShowModeEnabled' flag fixes the problem because
this flag is completely under the control of the app.