input1.asm: parseInputBuf(): restrict special interpretation of 'i' to solitary 'i'
Previously, the special interpretation would trigger if just the imaginary
component was empty. But that meant that '1i' would be parsed as '1i1' which
did not make sense. The new code triggers the special rule if *both* real and
imaginary pparts are empty. Only the solitary 'i' is special-cased to be '0i1'.
A solitary Angle or AngleDegree means '0i0' as before.