complex.asm: fix YToX and XRootY for real arguments within CRES mode
The real functions YToX() and XRootY() do not seem to behave well for negative
values of Y (as I had previously thought). So we use the same trick as the
SqRoot(), where if the numMode is CCRES and the arguments are real, we call
convert both arguments to complex, and call the complex versions CYToX() or
CXrootY(), then chop off the imaginary part if it's zero.

This causes a bit of weirdness in the results for negative values of Y and
integer values of X. For example, CubeRoot(-8) always returns -2 in both RRES
and CRES mode because it calls the result function XRootY() in both cases. But
XRootY(-8, 3) with an integer value '3' returns -2 in RRES mode, but 2 AngleDeg
60 (i.e. '1 i 1.73205081') in CRES mode. That's because the XRootY handler
calls XRootY() in RRES mode, but calls CXrootY() in CRES mode, and they return
different results. Both answers are mathematically correct, since there are 3
correct answers to the expression XRootY(-8, 3): 1 AngleDeg 60, 1 AngleDeg 180,
1 AngleDeg 300. Getting different results is a bit disconcerting if we don't
remember the underlying mathematics, but I'm not sure if it's worth the effort
to try to make the results the same.