argparser.asm: replace 'argModifierCanceled' with 'inputBufFlagsArgCancel' flag
Previously, the 'argModifier' was overloaded to hold the "Cancel" state. Moving
this to a separate flag cleans up the code for both the argModifier and
isCanceled handling.

Also lift ParseArgBuf() and displayStack() from arghandlers.asm to
argparser.asm. This improves the separation of concerns. The handlers simply
update buffers, sets flags, and so on. The parsing and updating the display is
now handled by the caller processArgCommands().