Update supported hardware: works on TI84P
I've been using this program on my TI84P extensively since November, so I can confirm that it does work. I think this amount of testing suffices.

Also, it was working on an older OS version too because I forgot to update it when I bought the calculator, but we'll just encourage latest ver.