menu.asm: extract altName, normalName, and nameSelector lookup from getMenuName() to extractMenuNames()
The getMenuName() routine invokes the menuNode.nameSelector() callback function
to select the normalName and altName. Therefore getMenuName() cannot be placed
in Flash Page 1, it must be on Flash Page 0 because that's where all the name
selector functions are. However, getMenuName() now calls extractMenuNames()
which depends only on the menuTable and does not invoke a callback. So
extractMenuNames() can be moved to Flash Page 1 eventually.

The getMenuName() routine also provides pointers to OP3 and OP4 as temporary
string buffers into extractMenuNames() which it fills those buffers with the
normal and alt names. This allows us to remove the menuStringBuf buffer from
the appBuffer area.