input1.asm: support entry of tagged records for date-like objects
Entry of the following tagged records are supported:

- D{y,m,d} - Date
- T{h,m,d} - Time
- DT{y,m,d,h,m,d} - DateTime
- TZ{tzh,tzm} - Offset
- DZ{y,m,d,h,m,s,tzh,tzm} - OffsetDateTime

Naked records (no record tagged, just the '{' and '}' braces) are still
supported as before. The record type is inferred from the number of fields
(i.e. number of commas). The Time record has the same number of fields as the
Date object, so a naked Time record is not supported. It must be entered as
"T{h,m,s}".