arg1.asm: refactor ParseArgBuf() to handle invalid arguments
Also handle varLetter arguments (A-Z,Theta) in the same way as numerical
arguments, so that 2 characters are always expected, instead of shortcircuiting
the input when a single letter is entered. This allows the A-Z,Theta letters to
be displayed on the screen, instead of instantly disappearing. The parser now
detects invalid combination of characters (letters combined with numbers) and
returns an 'argTypeInvalid' value for 'argType'. The processArgCommands()
routine detects that value, and throws an Err:Invalid.