varstack.asm,input.asm: replace inputBufFlagsClosedEmpty with rpnFlagsLiftEnabled
The parseAndClearInputBufEmpty() function sets rpnFlagsLiftEnabled instead of
inputBufFlagsClosedEmpty when the inputBuf is closed with an empty string. This
allows the various ReplaceXxx() and PushXxx() functions to behave correctly
regardless of whether the stack lift was disabled because of ENTER, or because
the inputBuf was closed with an empty string.

Fixes stack lift problems with PI, E and other functions which do not consume X
before producing the result. Previously, these functions ignored the stack lift
disable setting and pushed the duplicated X value before pushing the new
result. Now, if the stack lift is disabled, these functions overwrite the
existing X value, which matches the behavior of Classic HP RPN calculators.