handlers.asm: Fix STO regression, cannot store complex numbers and record objects.
Caused by 4a584f813 which implemented the feature to avoid termination of input
after an EXIT from a command argument prompt (e.g. STO, FIX, etc).
Unfortunately, moving the 'call closeInputAndRecallNone' to a later point in
the code caused the loss of A register to be lost, which identified the command
argument, which triggered an early return from the handler for complex and
record objects.